// src/js/model/game/utils/SServerUtil.ts
export const enum Agent2DData {
  STAMINA = 0,
  STAMINA_EFFORT = 1,
  STAMINA_RECOVERY = 2,
  STAMINA_CAPACITY = 3,

  VIEW_QUALITY = 4,
  VIEW_WIDTH = 5,

  FOCUS_SIDE = 6,
  FOCUS_UNUM = 7,

  KICK_COUNT = 8,
  DASH_COUNT = 9,
  TURN_COUNT = 10,
  CATCH_COUNT = 11,
  MOVE_COUNT = 12,
  TURN_NECK_COUNT = 13,
  VIEW_COUNT = 14,
  SAY_COUNT = 15,
  TACKLE_COUNT = 16,
  POINT_TO_COUNT = 17,
  ATTENTION_COUNT = 18
}
// ...
export const enum PlayerType2DParams {
  ID = 'id',
  PLAYER_SPEED_MAX = 'player_speed_max',
  STAMINA_INC_MAX = 'stamina_inc_max',
  PLAYER_DECAY = 'player_decay',
  INERTIA_MOMENT = 'inertia_moment',
  DASH_POWER_RATE = 'dash_power_rate',
  PLAYER_SIZE = 'player_size',
  KICKABLE_MARGIN = 'kickable_margin',
  KICK_RAND = 'kick_rand',
  EXTRA_STAMINA = 'extra_stamina',
  EFFORT_MAX = 'effort_max',
  EFFORT_MIN = 'effort_min',
  KICK_POWER_RATE = 'kick_power_rate',
  FOUL_DETECT_PROBABILITY = 'foul_detect_probability',
  CATCHABLE_AREA_L_STRETCH = 'catchable_area_l_stretch'
}